[RENAME/#30] 뷰모델 이름 변경
