+++ app/src/main/res/layout/fragment_actionlist.xml
@@ -82,7 +82,6 @@
         android:id="@+id/fcv_actionlist_main"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        android:layout_marginBottom="20dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
